use stored functions for photo data

+++ src/api/Models/Photos/UserRating.cs
@@ -3,6 +3,6 @@
 	public class UserRating
 	{
 		public int PhotoId {get;set;}
-		public byte Rating {get;set;}
+		public short Rating {get;set;}
 	}
 }
